Add Row type

+++ src/c++/Martian_Money/Source/bank.h
@@ -11,23 +11,22 @@ public:
 	Bank(Configuration Config);
 
 	// Facilitators
-	unsigned int Trade(unsigned int Offer_Type,unsigned int Offer_Adjective,
-			unsigned int Offer_Amount, unsigned int Receive_Type,
-			unsigned int Receive_Adjective);
+	size_t Trade(size_t Offer_Type, size_t Offer_Adjective, size_t Offer_Amount, size_t Receive_Type,
+			size_t Receive_Adjective);
 	void Make_Rates();
 
 	// Inspectors
 	double Get_Rate(unsigned int Type, unsigned int Adjective) const;
-	long Get_Money(unsigned int Type, unsigned int Adjective) const;
+	int Get_Money(unsigned int Type, unsigned int Adjective) const;
 	double Get_Normalizer() const;
 
 	// Mutators
-	void Set_Money(unsigned int Type, unsigned int Adjective, long Amount);
+	void Set_Money(unsigned int Type, unsigned int Adjective, int Amount);
 	void Set_Normalizer(double Normalizer);
 
 protected:
 	// Member Variables
-	Array<long> M_Money;
+	Array<int> M_Money;
 	Array<double> M_Rate;
 	double M_Normalizer;
 };
